requirements.txt
yfinance
pandas
scikit-learn
pytest
psycopg2-binary
